terminal_image_viewer/requirements.txt
imgcat
Pillow
Requests
tqdm
